+++ src/app/(private)/dashboard/@topScorers/page.tsx
@@ -1,8 +1,19 @@
+import { DashboardFacade } from "@/facade/dashboard/DashboardFacade";
 import TopScorersFt from "@/features/dashboard/infraestructure/TopScorersFt";
 import React from "react";
 
-function TopScorersSection() {
-  return <TopScorersFt />;
+async function TopScorersSection() {
+  const { players } = await getData();
+  return <TopScorersFt players={players} />;
 }
 
+async function getData() {
+  try {
+    const players = await DashboardFacade.getTopScorers();
+    return { players };
+  } catch (error) {
+    console.log(error);
+    return { players: [] };
+  }
+}
 export default TopScorersSection;
