+++ src/core/query/field-positions/fieldPositions.query.ts
@@ -2,7 +2,9 @@ import { getFieldPositions } from "@/services/player/program/fieldPositions/fiel
 import { queryOptions } from "@tanstack/react-query";
 import { Effect } from "effect";
 
-export const fieldPositionsQueryOptions = (access_token: string | undefined) =>
+export const fieldPositionsQueryOptions = (
+  access_token: string | undefined | null
+) =>
   queryOptions({
     queryKey: ["field-positions"],
     queryFn: () => Effect.runPromise(getFieldPositions(access_token!)),
