+++ src/features/dashboard/infraestructure/TopAssitsFt.tsx
@@ -1,14 +1,16 @@
 import TopAssists from "@/components/Dashboard/TopAsists/TopAsists";
-import { DashboardFacade } from "@/facade/dashboard/DashboardFacade";
+import { FulfilledMatchAparition } from "@/features/aparition/domain/aparition.schema";
 
-const LIMIT = 5;
-async function TopAssistsFt() {
-  const { players } = await getData();
+async function TopAssistsFt({
+  players,
+}: {
+  players: FulfilledMatchAparition[];
+}) {
   return (
     <div className="p-2">
       <div className="px-5 my-2">
         <h3 className="text-lg font-bold tracking-tight text-white max-lg:text-center">
-          Top {LIMIT} Assists
+          Top 5 Assists
         </h3>
       </div>
       <div className="px-5">
@@ -16,11 +18,6 @@ async function TopAssistsFt() {
       </div>
     </div>
   );
-}
-
-async function getData() {
-  const players = await DashboardFacade.getTopAsists();
-  return { players };
 }
 
 export default TopAssistsFt;
